% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gttable1.R
\name{gttable1}
\alias{gttable1}
\title{Create table1}
\usage{
gttable1(
  data = NULL,
  group_var = NULL,
  cat_var = NULL,
  cont_var = NULL,
  out_format = "gtobj",
  digits = 2,
  pDigits = 3,
  method = "auto",
  ...
)
}
\arguments{
\item{data}{A dataframe.}

\item{group_var}{A character, a column name (quoted) in data. Summary statistics will be calculated 
separately for each level of the by variable. If NULL, summary statistics are calculated using 
all observations.}

\item{cat_var}{A vector, column name (quoted) in data. Indicative categorical variables. 
If NULL, all categorical variables will be calculated.}

\item{cont_var}{A vector, column name (quoted) in data. Indicative continuous variables.
If NULL, all continuous variables will be calculated.}

\item{out_format}{A character, the format of output (e.g. "gtobj", "dataframe"). Default is "gtobj".}

\item{digits}{A numeric, number of decimal places of continuous variables.}

\item{pDigits}{A numeric, number of decimal places of p value.}

\item{method}{character, methods for gttable1 (e.g. "auto","define"). Default is "auto".
auto : equal to gtsummary package; define : automatically choose appropriate statistical methods 
according to normality and homogeneity of variance.}

\item{...}{reference gtsummary::tbl_summary.}
}
\value{
Return a gt object or dataframe.
}
\description{
\code{gttable1} is a function to create table1
}
\examples{
data(pbc, package = "survival")
cont_var <- c("bili","chol","copper","alk.phos","trig")
cat_var <- c("sex","status","stage")
group_var <- "trt"
gttable1(data=pbc,out_format="dataframe",pDigits=3,method="auto")
gttable1(data=pbc,cat_var=cat_var,out_format="gtobj",pDigits=3,method="define")
gttable1(data=pbc,group_var=group_var,cat_var=cat_var,cont_var=cont_var,out_format="gtobj",pDigits=3,method="auto")
gttable1(data=pbc,group_var=group_var,cat_var=cat_var,cont_var=cont_var,out_format="gtobj",pDigits=3,method="auto",missing="always")

}
